@@ -124,8 +124,8 @@ RCT_NOT_IMPLEMENTED(- (instancetype)init)
   // Do a linear remap of the two frames to safeguard against variable framerates
   NSNumber *fromFrameValue = _frames[startIndex];
   NSNumber *toFrameValue = _frames[nextIndex];
-  NSTimeInterval fromInterval = startIndex * RCTSingleFrameInterval;
-  NSTimeInterval toInterval = nextIndex * RCTSingleFrameInterval;
+  NSTimeInterval fromInterval = (double)startIndex * RCTSingleFrameInterval;
+  NSTimeInterval toInterval = (double)nextIndex * RCTSingleFrameInterval;
 
   // Interpolate between the individual frames to ensure the animations are
   //smooth and of the proper duration regardless of the framerate.
