@@ -28,6 +28,7 @@ using namespace facebook::react;
   NSDictionary *_currentInterfaceDimensions;
 }
 
+@synthesize bridge = _bridge;
 @synthesize moduleRegistry = _moduleRegistry;
 
 RCT_EXPORT_MODULE()
@@ -56,7 +57,7 @@ RCT_EXPORT_MODULE()
                                                name:UIApplicationDidChangeStatusBarOrientationNotification
                                              object:nil];
 
-  _currentInterfaceDimensions = RCTExportedDimensions(_moduleRegistry);
+  _currentInterfaceDimensions = RCTExportedDimensions(_moduleRegistry, _bridge);
 
   [[NSNotificationCenter defaultCenter] addObserver:self
                                            selector:@selector(interfaceFrameDidChange)
@@ -67,14 +68,6 @@ RCT_EXPORT_MODULE()
                                            selector:@selector(interfaceFrameDidChange)
                                                name:RCTUserInterfaceStyleDidChangeNotification
                                              object:nil];
-}
-
-- (void)setModuleRegistry:(RCTModuleRegistry *)moduleRegistry
-{
-  if (!moduleRegistry) {
-    RCTAssert(false, @"moduleRegistry should not be nil, _moduleRegistry is nil: %d", _moduleRegistry == nil);
-  }
-  _moduleRegistry = moduleRegistry;
 }
 
 static BOOL RCTIsIPhoneX()
@@ -103,7 +96,7 @@ static BOOL RCTIsIPhoneX()
   return isIPhoneX;
 }
 
-static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
+static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry, RCTBridge *bridge)
 {
   RCTAssertMainQueue();
   RCTDimensions dimensions;
@@ -112,7 +105,7 @@ static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
         (RCTAccessibilityManager *)[moduleRegistry moduleForName:"AccessibilityManager"];
     dimensions = RCTGetDimensions(accessibilityManager ? accessibilityManager.multiplier : 1.0);
   } else {
-    RCTAssert(false, @"ModuleRegistry must be set to properly init dimensions.");
+    RCTAssert(false, @"ModuleRegistry must be set to properly init dimensions. Bridge exists: %d", bridge != nil);
   }
   __typeof(dimensions.window) window = dimensions.window;
   NSDictionary<NSString *, NSNumber *> *dimsWindow = @{
@@ -140,9 +133,10 @@ static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
 {
   __block NSDictionary<NSString *, id> *constants;
   RCTModuleRegistry *moduleRegistry = _moduleRegistry;
+  RCTBridge *bridge = _bridge;
   RCTUnsafeExecuteOnMainQueueSync(^{
     constants = @{
-      @"Dimensions" : RCTExportedDimensions(moduleRegistry),
+      @"Dimensions" : RCTExportedDimensions(moduleRegistry, bridge),
       // Note:
       // This prop is deprecated and will be removed in a future release.
       // Please use this only for a quick and temporary solution.
@@ -157,12 +151,14 @@ static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
 - (void)didReceiveNewContentSizeMultiplier
 {
   RCTModuleRegistry *moduleRegistry = _moduleRegistry;
+  RCTBridge *bridge = _bridge;
   RCTExecuteOnMainQueue(^{
   // Report the event across the bridge.
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    [[moduleRegistry moduleForName:"EventDispatcher"] sendDeviceEventWithName:@"didUpdateDimensions"
-                                                                         body:RCTExportedDimensions(moduleRegistry)];
+    [[moduleRegistry moduleForName:"EventDispatcher"]
+        sendDeviceEventWithName:@"didUpdateDimensions"
+                           body:RCTExportedDimensions(moduleRegistry, bridge)];
 #pragma clang diagnostic pop
   });
 }
@@ -186,8 +182,9 @@ static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
        !UIInterfaceOrientationIsLandscape(nextOrientation))) {
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    [[_moduleRegistry moduleForName:"EventDispatcher"] sendDeviceEventWithName:@"didUpdateDimensions"
-                                                                          body:RCTExportedDimensions(_moduleRegistry)];
+    [[_moduleRegistry moduleForName:"EventDispatcher"]
+        sendDeviceEventWithName:@"didUpdateDimensions"
+                           body:RCTExportedDimensions(_moduleRegistry, _bridge)];
 #pragma clang diagnostic pop
   }
 
@@ -204,7 +201,7 @@ static NSDictionary *RCTExportedDimensions(RCTModuleRegistry *moduleRegistry)
 
 - (void)_interfaceFrameDidChange
 {
-  NSDictionary *nextInterfaceDimensions = RCTExportedDimensions(_moduleRegistry);
+  NSDictionary *nextInterfaceDimensions = RCTExportedDimensions(_moduleRegistry, _bridge);
 
   if (!([nextInterfaceDimensions isEqual:_currentInterfaceDimensions])) {
 #pragma clang diagnostic push
