@@ -26,15 +26,21 @@ void setLocalizationLanguagePack(NSDictionary<NSString *, NSString *> *pack)
 
 + (NSString *)RCTLocalizedString:(NSString *)oldString withDescription:(NSString *)description
 {
+  NSString *candidate = nil;
+
   if (_delegate != nil) {
-    return [_delegate localizedString:oldString withDescription:description];
+    candidate = [_delegate localizedString:oldString withDescription:description];
   }
 
-  if (_languagePack != nil) {
-    return _languagePack[oldString];
+  if (candidate == nil && _languagePack != nil) {
+    candidate = _languagePack[oldString];
   }
 
-  return oldString;
+  if (candidate == nil) {
+    candidate = oldString;
+  }
+
+  return candidate;
 }
 
 @end
