@@ -30,7 +30,7 @@ babelRegisterOnly([]);
 
 const transformer = require('metro-react-native-babel-transformer');
 module.exports = {
-  process(src /*: string */, file /*: string */) /*: string */ {
+  process(src /*: string */, file /*: string */) /*: {code: string, ...} */ {
     if (nodeFiles.test(file)) {
       // node specific transforms only
       return babelTransformSync(src, {
@@ -38,7 +38,7 @@ module.exports = {
         sourceType: 'script',
         ...nodeOptions,
         ast: false,
-      }).code;
+      });
     }
 
     const {ast} = transformer.transform({
@@ -111,7 +111,7 @@ module.exports = {
         sourceMaps: true,
       },
       src,
-    ).code;
+    );
   },
 
   getCacheKey: (createCacheKeyFunction([
