@@ -23,29 +23,29 @@ import RNTesterNavBar, {navBarHeight} from './components/RNTesterNavbar';
 import RNTesterList from './utils/RNTesterList';
 import {
   Screens,
-  initialState,
+  initialNavState,
   getExamplesListWithBookmarksAndRecentlyUsed,
-  getInitialStateFromAsyncStorage,
 } from './utils/testerStateUtils';
-import {useAsyncStorageReducer} from './utils/useAsyncStorageReducer';
-import {RNTesterReducer, RNTesterActionsType} from './utils/RNTesterReducer';
+import {
+  RNTesterNavReducer,
+  RNTesterNavActionsType,
+} from './utils/RNTesterNavReducer';
 import {RNTesterThemeContext, themes} from './components/RNTesterTheme';
 import RNTTitleBar from './components/RNTTitleBar';
 import {RNTesterEmptyBookmarksState} from './components/RNTesterEmptyBookmarksState';
+import {RNTesterJsStallsProvider} from './utils/RNTesterJsStallsContext';
 
-const APP_STATE_KEY = 'RNTesterAppState.v3';
-
-// RNTester App currently uses AsyncStorage from react-native for storing navigation state
+// RNTester App currently uses in memory storage for storing navigation state
 // and bookmark items.
-// TODO: Vendor AsyncStorage or create our own.
+// TODO: Identify/implement solution for async device storage.
 LogBox.ignoreLogs([/AsyncStorage has been extracted from react-native/]);
 
 const RNTesterApp = (): React.Node => {
-  const [state, dispatch] = useAsyncStorageReducer(
-    RNTesterReducer,
-    initialState,
-    APP_STATE_KEY,
+  const [state, dispatch] = React.useReducer<Function, Object>(
+    RNTesterNavReducer,
+    initialNavState,
   );
+
   const colorScheme = useColorScheme();
 
   const {
@@ -57,17 +57,6 @@ const RNTesterApp = (): React.Node => {
     recentlyUsed,
   } = state;
 
-  React.useEffect(() => {
-    getInitialStateFromAsyncStorage(APP_STATE_KEY).then(
-      initialStateFromStorage => {
-        dispatch({
-          type: RNTesterActionsType.INIT_FROM_STORAGE,
-          data: initialStateFromStorage,
-        });
-      },
-    );
-  }, [dispatch]);
-
   const examplesList = React.useMemo(
     () =>
       getExamplesListWithBookmarksAndRecentlyUsed({bookmarks, recentlyUsed}),
@@ -76,7 +65,7 @@ const RNTesterApp = (): React.Node => {
 
   const handleBackPress = React.useCallback(() => {
     if (activeModuleKey != null) {
-      dispatch({type: RNTesterActionsType.BACK_BUTTON_PRESS});
+      dispatch({type: RNTesterNavActionsType.BACK_BUTTON_PRESS});
     }
   }, [dispatch, activeModuleKey]);
 
@@ -103,7 +92,7 @@ const RNTesterApp = (): React.Node => {
   const handleModuleCardPress = React.useCallback(
     ({exampleType, key, title}) => {
       dispatch({
-        type: RNTesterActionsType.MODULE_CARD_PRESS,
+        type: RNTesterNavActionsType.MODULE_CARD_PRESS,
         data: {exampleType, key, title},
       });
     },
@@ -113,7 +102,7 @@ const RNTesterApp = (): React.Node => {
   const handleModuleExampleCardPress = React.useCallback(
     exampleName => {
       dispatch({
-        type: RNTesterActionsType.EXAMPLE_CARD_PRESS,
+        type: RNTesterNavActionsType.EXAMPLE_CARD_PRESS,
         data: {key: exampleName},
       });
     },
@@ -123,7 +112,7 @@ const RNTesterApp = (): React.Node => {
   const toggleBookmark = React.useCallback(
     ({exampleType, key}) => {
       dispatch({
-        type: RNTesterActionsType.BOOKMARK_PRESS,
+        type: RNTesterNavActionsType.BOOKMARK_PRESS,
         data: {exampleType, key},
       });
     },
@@ -133,7 +122,7 @@ const RNTesterApp = (): React.Node => {
   const handleNavBarPress = React.useCallback(
     args => {
       dispatch({
-        type: RNTesterActionsType.NAVBAR_PRESS,
+        type: RNTesterNavActionsType.NAVBAR_PRESS,
         data: {screen: args.screen},
       });
     },
@@ -170,40 +159,42 @@ const RNTesterApp = (): React.Node => {
 
   return (
     <RNTesterThemeContext.Provider value={theme}>
-      <RNTTitleBar
-        title={title}
-        theme={theme}
-        onBack={activeModule ? handleBackPress : null}
-        documentationURL={activeModule?.documentationURL}
-      />
-      <View
-        style={StyleSheet.compose(styles.container, {
-          backgroundColor: theme.GroupedBackgroundColor,
-        })}>
-        {activeModule != null ? (
-          <RNTesterModuleContainer
-            module={activeModule}
-            example={activeModuleExample}
-            onExampleCardPress={handleModuleExampleCardPress}
-          />
-        ) : screen === Screens.BOOKMARKS &&
-          examplesList.bookmarks.length === 0 ? (
-          <RNTesterEmptyBookmarksState />
-        ) : (
-          <RNTesterModuleList
-            sections={activeExampleList}
-            toggleBookmark={toggleBookmark}
-            handleModuleCardPress={handleModuleCardPress}
-          />
-        )}
-      </View>
-      <View style={styles.bottomNavbar}>
-        <RNTesterNavBar
-          screen={screen || Screens.COMPONENTS}
-          isExamplePageOpen={!!activeModule}
-          handleNavBarPress={handleNavBarPress}
+      <RNTesterJsStallsProvider>
+        <RNTTitleBar
+          title={title}
+          theme={theme}
+          onBack={activeModule ? handleBackPress : null}
+          documentationURL={activeModule?.documentationURL}
         />
-      </View>
+        <View
+          style={StyleSheet.compose(styles.container, {
+            backgroundColor: theme.GroupedBackgroundColor,
+          })}>
+          {activeModule != null ? (
+            <RNTesterModuleContainer
+              module={activeModule}
+              example={activeModuleExample}
+              onExampleCardPress={handleModuleExampleCardPress}
+            />
+          ) : screen === Screens.BOOKMARKS &&
+            examplesList.bookmarks.length === 0 ? (
+            <RNTesterEmptyBookmarksState />
+          ) : (
+            <RNTesterModuleList
+              sections={activeExampleList}
+              toggleBookmark={toggleBookmark}
+              handleModuleCardPress={handleModuleCardPress}
+            />
+          )}
+        </View>
+        <View style={styles.bottomNavbar}>
+          <RNTesterNavBar
+            screen={screen || Screens.COMPONENTS}
+            isExamplePageOpen={!!activeModule}
+            handleNavBarPress={handleNavBarPress}
+          />
+        </View>
+      </RNTesterJsStallsProvider>
     </RNTesterThemeContext.Provider>
   );
 };
