@@ -140,9 +140,12 @@ class MultiColumnExample extends React.PureComponent<
       getItemLayout(data, index).length + 2 * (CARD_MARGIN + BORDER_WIDTH);
     return {length, offset: length * index, index};
   }
-  _renderItemComponent = ({item}: RenderItemProps<any | Item>) => {
+  _renderItemComponent = ({item, accessibilityCollectionItem}) => {
     return (
-      <View style={styles.card}>
+      <View
+        importantForAccessibility="yes"
+        accessibilityCollectionItem={accessibilityCollectionItem}
+        style={styles.card}>
         <ItemComponent
           item={item}
           fixedHeight={this.state.fixedHeight}
