@@ -158,74 +158,45 @@ class LoopExample extends React.Component<{...}, $FlowFixMeState> {
   }
 }
 
-const RNTesterSettingSwitchRow = require('../../components/RNTesterSettingSwitchRow');
-class InternalSettings extends React.Component<
-  {...},
-  {
-    busyTime: number | string,
-    filteredStall: number,
-    ...
-  },
-> {
-  _stallInterval: ?number;
-  render() {
-    return (
-      <View>
-        <RNTesterSettingSwitchRow
-          initialValue={false}
-          label="Force JS Stalls"
-          onEnable={() => {
-            /* $FlowFixMe[incompatible-type] (>=0.63.0 site=react_native_fb)
-             * This comment suppresses an error found when Flow v0.63 was
-             * deployed. To see the error delete this comment and run Flow. */
-            this._stallInterval = setInterval(() => {
-              const start = Date.now();
-              console.warn('burn CPU');
-              while (Date.now() - start < 100) {}
-            }, 300);
-          }}
-          onDisable={() => {
-            /* $FlowFixMe[incompatible-call] (>=0.63.0 site=react_native_fb)
-             * This comment suppresses an error found when Flow v0.63 was
-             * deployed. To see the error delete this comment and run Flow. */
-            clearInterval(this._stallInterval || 0);
-          }}
-        />
-        <RNTesterSettingSwitchRow
-          initialValue={false}
-          label="Track JS Stalls"
-          onEnable={() => {
-            require('react-native/Libraries/Interaction/JSEventLoopWatchdog').install(
-              {
-                thresholdMS: 25,
-              },
-            );
-            this.setState({busyTime: '<none>'});
-            require('react-native/Libraries/Interaction/JSEventLoopWatchdog').addHandler(
-              {
-                onStall: ({busyTime}) =>
-                  this.setState(state => ({
-                    busyTime,
-                    filteredStall:
-                      (state.filteredStall || 0) * 0.97 + busyTime * 0.03,
-                  })),
-              },
-            );
-          }}
-          onDisable={() => {
-            console.warn('Cannot disable yet....');
-          }}
-        />
-        {this.state && (
-          <Text>
-            {`JS Stall filtered: ${Math.round(this.state.filteredStall)}, `}
-            {`last: ${this.state.busyTime}`}
-          </Text>
-        )}
-      </View>
-    );
-  }
-}
+const RNTesterSettingSwitchRow =
+  require('../../components/RNTesterSettingSwitchRow').RNTesterSettingSwitchRow;
+const RNTesterJsStallsContext = require('../../utils/RNTesterJsStallsContext');
+
+const InternalSettings = () => {
+  const {
+    state,
+    onDisableForceJsStalls,
+    onEnableForceJsStalls,
+    onEnableJsStallsTracking,
+    onDisableJsStallsTracking,
+  } = React.useContext(RNTesterJsStallsContext.RNTesterJsStallsContext);
+
+  const {stallInterval, busyTime, filteredStall, tracking} = state;
+  return (
+    <View>
+      <RNTesterSettingSwitchRow
+        initialValue={false}
+        label="Force JS Stalls"
+        active={stallInterval !== null}
+        onEnable={onEnableForceJsStalls}
+        onDisable={onDisableForceJsStalls}
+      />
+      <RNTesterSettingSwitchRow
+        initialValue={false}
+        label="Track JS Stalls"
+        active={tracking}
+        onEnable={onEnableJsStallsTracking}
+        onDisable={onDisableJsStallsTracking}
+      />
+      {tracking && (
+        <Text>
+          {`JS Stall filtered: ${Math.round(filteredStall)}, `}
+          {`last: ${busyTime !== null ? busyTime.toFixed(8) : '<none>'}`}
+        </Text>
+      )}
+    </View>
+  );
+};
 
 class EventExample extends React.Component<{...}, $FlowFixMeState> {
   state = {
