@@ -16,12 +16,10 @@ module.exports = {
   commands: [...ios.commands, ...android.commands],
   platforms: {
     ios: {
-      linkConfig: ios.linkConfig,
       projectConfig: ios.projectConfig,
       dependencyConfig: ios.dependencyConfig,
     },
     android: {
-      linkConfig: android.linkConfig,
       projectConfig: android.projectConfig,
       dependencyConfig: android.dependencyConfig,
     },
@@ -32,7 +30,7 @@ module.exports = {
   reactNativePath: '.',
   project: {
     ios: {
-      project: './packages/rn-tester/RNTesterPods.xcworkspace',
+      sourceDir: './packages/rn-tester',
     },
     android: {
       sourceDir: './packages/rn-tester',
